Initial test release
This is my first commit to this project! However, the code won't compile
for me, not sure if this is because of MinGW32 being unfriendly or there's
something wrong with the code itself. It gives me a common.c:32 error,
even though as you can see I didn't even change that file. Any insight on
this would be useful.

As for actual changes, I've changed the Turn function so it wouldn't have
the sound parameter in Cards.Turn() at all, and instead I've set up
SOUND_PLAY() there. Though the best, mod-compatible way of writing card
configurations would be to create functions for each of the actions
(damage, adding/taking resources and mines etc.) and later on setting up a
kind of an array/struct to hold all the info on cards available to the
game.

Signed-off-by: GreatEmerald <pastas4@gmail.com>