Configuration changes
The code is now more configuration-compatible: added tower, wall, resource
generator levels and resource quantities needed to win the game to the
global variable list, added support to up to 50 "cursed" cards like
LodeStone, added resource victory support and a config variable for winning
by getting all resources up to a specified level (as in the original) or
only one (more likely to be accomplished). The dialog boxes now state the
reason for winning/losing.

If wall levels are equal, Flood Water now affects the enemy instead of the
player which used the card, as in the original.

Added original Arcomage menu as a config option (off by default). The font
of the win/lose dialog box is now smaller to have enough space for the
extended description (ideally the font should remain big but be in two
lines, though).

Signed-off-by: GreatEmerald <pastas4@gmail.com>